refactor: 收口 Bundle 领域边界
