+++ xl_updata_tool/app/features/importer/processing.py
@@ -17,7 +17,7 @@ import time
 from app.core.logger import logger, timed
 from app.core.task_context import stage_operation, task_operation
 from app.core.path_utils import get_base_dir, get_tools_dir
-from app.core.bundle_parser import fix_bundle_inplace
+from app.platform.bundle_parser import fix_bundle_inplace
 from app.core.file_utils import replace_directory
 from app.core.lua_repository import cleanup_lua_staging, publish_lua_version
 from app.ui.workers.lua_decrypt import decompile_lua_dir
