+++ xl_updata_tool/docs/worklog/issue-50-architecture-finalization/progress.md
@@ -16,7 +16,7 @@
 - [completed] P5：迁移 Importer / AssetStudio Worker；Qt-free 导入处理进入 Feature，生产 Feature Worker 不再依赖旧 `app/ui/workers/import_as.py`，旧路径保留薄兼容门面。
 - [completed] P6a：归位音频与预览领域实现；`app/core` 对应路径降为兼容转发，Feature 内部不再从 core 获取这些领域实现。
 - [completed] P6b-版本：归位版本仓库、差异、下载计划、更新持久化、本地 Bundle 同步和种子实现；对应 core 路径降为兼容转发。
-- [pending] P6b-Bundle：归位 Bundle Parser/Selector/Manager，并收口 Versions/Importer 调用边界。
+- [completed] P6b-Bundle：归位 Bundle Parser/Selector/Manager，并收口 Versions/Importer 调用边界；旧 core 路径保留兼容转发。
 - [pending] P6c：归位角色领域仓库/展示实现。
 - [pending] P6d：收口平台与诊断基础设施。
